TICKET-2.3: Controller all good for back and save file. FXML files updated for both lesson_generator-teacher.fxml and lesson_plan-teacher.fxml - they both are stretchable with expanding text fields
